Moved Leaderboard API functions to profile
