+++ HackIllinois/DataSources/HILeaderboardDataSource.swift
@@ -1,0 +1,20 @@
+//
+//  HIProfileDataSource.swift
+//  HackIllinois
+//
+//  Created by HackIllinois Team on 11/16/21.
+//  Copyright © 2021 HackIllinois. All rights reserved.
+//  This file is part of the Hackillinois iOS App.
+//  The Hackillinois iOS App is open source software, released under the University of
+//  Illinois/NCSA Open Source License. You should have received a copy of
+//  this license in a file with the distribution.
+//
+
+import Foundation
+import CoreData
+import HIAPI
+import os
+
+final class HILeaderboardDataSource {
+    
+}
